backend/requirements.txt
fastapi
uvicorn[standard]
wandb
openai
python-dotenv
pyyaml
